requirements.txt
# Core ML
torch==2.2.0
pytorch-lightning==2.2.0
torch-geometric==2.5.2

# Chemistry
rdkit==2024.3.1
biopython==1.83

# RL / stats
scipy==1.12.0
numpy==1.26.4
pandas==2.2.1

# DB / MLflow
sqlalchemy==2.0.29
pyodbc==5.1.0
mlflow==2.12.1

# UI
streamlit==1.33.0
altair==5.3.0
Pillow==10.3.0

# Notebook execution
jupyter==1.0.0
nbconvert==7.16.3
ipykernel==6.29.4
